refactor: add new world structure
